Fix cookieless modes in IE
