+++ src/js/out_queue.js
@@ -382,6 +382,8 @@ export function OutQueueManager(
           executeQueue();
         }
       }, connectionTimeout);
+    } else {
+      executingQueue = false;
     }
   }
 
@@ -447,7 +449,7 @@ export function OutQueueManager(
 
   return {
     enqueueRequest: enqueueRequest,
-    executeQueue: () => { if (!executingQueue) executeQueue(); },
+    executeQueue: () => { if (!executingQueue) { executeQueue(); } },
     setUseLocalStorage: (localStorage) => { useLocalStorage = localStorage; },
     setAnonymousTracking: (anonymous) => { anonymousTracking = anonymous; }
   };
